Crate sfml

Rust bindings for SFML, the Simple and Fast Multimedia Library.

Prerequisites

Rust 1.36

SFML 2.5 and CSFML 2.5 must be installed on your computer. You can download them here:

SFML 2.5: http://www.sfml-dev.org/download.php

CSFML 2.5: http://www.sfml-dev.org/download/csfml/

Supported platforms:

Linux

Windows

Mac OS X

License

This software is a binding of the SFML library created by Laurent Gomila, which is provided under the Zlib/png license.

This software is provided under the same license than the SFML, the Zlib/png license.

Modules

audio
Sounds, streaming (musics or custom sources), recording, spatialization

graphics
2D graphics module: sprites, text, shapes..

system
Base module of SFML, defining various utilities.

window
Provides OpenGL-based windows, and abstractions for events and input handling.