Module sfml::audio

Sounds, streaming (musics or custom sources), recording, spatialization

Re-exports

pub use self::capture::SoundBufferRecorder;

pub use self::capture::SoundRecorder;

pub use self::capture::SoundRecorderDriver;

Modules

capture
Types and helper functions dealing with audio capture.

listener
The audio listener is the point in the scene from where all the sounds are heard.

Structs

Music
Streamed music played from an audio file.

Sound
Regular sound that can be played in the audio environment.

SoundBuffer
Storage for audio samples defining a sound.

SoundStreamPlayer
Player for custom streamed audio sources. See SoundStream.

TimeSpan
Structure defining a time range

Enums

SoundStatus
Enumeration of statuses for sounds and musics

Traits

SoundSource
Base trait defining a sound's properties.

SoundStream
Trait for streamed audio sources.