Module sfml::audio::capture

Types and helper functions dealing with audio capture.

Structs

SetDeviceError
Error trying to set a capture device.

SoundBufferRecorder
Specialized SoundRecorder which stores the captured audio data into a sound buffer.

SoundRecorderDriver
Type that "drives" custom sound recorders.

Traits

SoundRecorder
Trait for processing captured sound data.

Functions

available_devices
Get a list of the names of all available audio capture devices.

default_device
Get the name of the default audio capture device.

is_available
Check if the system supports audio capture