Struct sfml::audio::capture::SoundRecorderDriver

pub struct SoundRecorderDriver<'a, R: 'a> { /* fields omitted */ }

Type that "drives" custom sound recorders.

It does the actual recording, and feeds the custom sound recorder with the recorded data.

Implementations

impl<'a, R: SoundRecorder> SoundRecorderDriver<'a, R>

pub fn new(sound_recorder: &'a mut R) -> Self

Creates a new SoundRecorderDriver with the specified SoundRecorder.

pub fn start(&mut self, sample_rate: u32) -> bool

Start the capture.

The sample_rate parameter defines the number of audio samples captured per second. The higher, the better the quality (for example, 44100 samples/sec is CD quality). This function uses its own thread so that it doesn't block the rest of the program while the capture runs. Please note that only one capture can happen at the same time. You can select which capture device will be used, by passing the name to the set_device method. If none was selected before, the default capture device will be used. You can get a list of the names of all available capture devices by calling available_devices.

Returns whether the start of capture was successful.

pub fn stop(&mut self) -> &mut R

Stop the capture, lending out the underlying SoundRecorder.

pub fn sample_rate(&self) -> u32

Get the sample rate.

The sample rate defines the number of audio samples captured per second. The higher, the better the quality (for example, 44100 samples/sec is CD quality).

pub fn set_channel_count(&mut self, channel_count: u32)

Set the channel count of the audio capture device.

This method allows you to specify the number of channels used for recording. Currently only 16-bit mono and 16-bit stereo are supported.

Parameters

channel_count Number of channels. Currently only mono (1) and stereo (2) are supported.

pub fn channel_count(&self) -> u32

Get the number of channels used by this recorder.

Currently only mono and stereo are supported, so the value is either 1 (for mono) or 2 (for stereo).

pub fn set_processing_interval(&mut self, interval: Time)

Set the processing interval.

The processing interval controls the period between calls to SoundRecorder::on_process_samples. You may want to use a small interval if you want to process the recorded data in real time, for example.

Note: this is only a hint, the actual period may vary. So don't rely on this parameter to implement precise timing.

The default processing interval is 100 ms.

pub fn device(&self) -> String

Get the name of the current audio capture device.

pub fn set_device(&mut self, name: &str) -> Result<(), SetDeviceError>

Set the audio capture device.

This function sets the audio capture device to the device with the given name. It can be called on the fly (i.e: while recording). If you do so while recording and opening the device fails, it stops the recording.