Struct sfml::audio::Sound

pub struct Sound<'s> { /* fields omitted */ }

Regular sound that can be played in the audio environment.

Sound is the type to use to play sounds.

It provides:

Control (play, pause, stop)

Ability to modify output parameters in real-time (pitch, volume, ...)

3D spatial features (position, attenuation, ...).

Sound is perfect for playing short sounds that can fit in memory and require no latency, like foot steps or gun shots. For longer sounds, like background musics or long speeches, rather see Music (which is based on streaming).

In order to work, a sound must be given a buffer of audio data to play. Audio data (samples) is stored in SoundBuffer, and attached to a sound with the set_buffer function. The buffer object attached to a sound must remain alive as long as the sound uses it. Note that multiple sounds can use the same sound buffer at the same time.

Usage example

use sfml::audio::{Sound, SoundBuffer};

let buffer = SoundBuffer::from_file("sound.wav").unwrap();
let mut sound = Sound::with_buffer(&buffer);
sound.play();

Implementations

impl<'s> Sound<'s>

pub fn new() -> Sound<'s>

Create a new Sound

pub fn with_buffer(buffer: &SoundBuffer) -> Sound<'_>

Create a new Sound with a buffer

pub fn set_looping(&mut self, looping: bool)

Sets whether this sound should loop or not.

pub fn is_looping(&self) -> bool

Tell whether or not a sound is in loop mode

Return true if the sound is looping, false otherwise

pub fn play(&mut self)

Start or resume playing a sound

This function starts the sound if it was stopped, resumes it if it was paused, and restarts it from beginning if it was it already playing. This function uses its own thread so that it doesn't block the rest of the program while the sound is played.

pub fn pause(&mut self)

Pause a sound

This function pauses the sound if it was playing, otherwise (sound already paused or stopped) it has no effect.

pub fn stop(&mut self)

Stop playing a sound

This function stops the sound if it was playing or paused, and does nothing if it was already stopped. It also resets the playing position (unlike pause).

pub fn status(&self) -> SoundStatus

Get the current status of a sound (stopped, paused, playing)

Return current status

pub fn playing_offset(&self) -> Time

Get the current playing position of a sound

Return the current playing position

pub fn set_playing_offset(&mut self, time_offset: Time)

Change the current playing position of a sound

The playing position can be changed when the sound is either paused or playing.

Arguments

timeOffset - New playing position

pub fn set_buffer(&mut self, buffer: &'s SoundBuffer)

Set the source buffer containing the audio data to play

Arguments

buffer - Sound buffer to attach to the sound

pub fn buffer(&self) -> Option<&SoundBuffer>

Get the audio buffer attached to a sound

Return an option to Sound buffer attached to the sound or None