Module sfml::graphics

2D graphics module: sprites, text, shapes..

Re-exports

pub use self::blend_mode::BlendMode;

Modules

blend_mode
Available blending modes for drawing

glsl
GLSL types.

Structs

CircleShape
Specialized shape representing a circle.

Color
Utility type for manpulating RGBA colors

ConvexShape
Specialized shape representing a convex polygon

ConvexShapePoints
An iterator over the points of a ConvexShape.

CustomShape
A custom textured shape with outline.

Font
Type for loading and manipulating character fonts.

FontInfo
Holds various information about a font.

Glyph
Structure describing a glyph.

Image
Loading, manipulating and saving images.

Rect
Utility type for manipulating 2D axis-aligned rectangles.

RectangleShape
Specialized shape representing a rectangle

RenderStates
Define the states used for drawing to a RenderTarget.

RenderTexture
Target for off-screen 2D rendering into a texture

RenderWindow
Window that can serve as a target for 2D drawing.

Shader
Shader type (vertex, geometry and fragment).

Sprite
Drawable representation of a texture

Text
Graphical text

TextStyle
Available text styles.

Texture
Image living on the graphics card that can be used for drawing.

Transform
Define a 3x3 transform matrix.

Vertex
Define a point with color and texture coordinates.

VertexArray
Define a set of one or more 2D primitives

Vertices
An iterator over the vertice of a VertexArray.

View
2D camera that defines what region is shown on screen

Enums

PrimitiveType
Types of primitives that a VertexArray can render.

Traits

CustomShapePoints
The points of a custom shape.

Drawable
The trait drawable is inherited by each object who can be drawn in a RenderTarget

RenderTarget
Trait for all render targets (window, texture, ...)

Shape
Trait for textured shapes with outline.

Transformable
Decomposed transform defined by a position, a rotation and a scale.

Type Definitions

FloatRect
A Rect of f32.

IntRect
A Rect of i32.