Struct sfml::graphics::Font

#[repr(C)]pub struct Font { /* fields omitted */ }

Type for loading and manipulating character fonts.

Fonts can be loaded from a file, from memory or from a custom stream, and supports the most common types of fonts.

See the from_file function for the complete list of supported formats.

Once it is loaded, a Font instance provides three types of information about the font:

Global metrics, such as the line spacing

Per-glyph metrics, such as bounding box or kerning

Pixel representation of glyphs

Fonts alone are not very useful: they hold the font data but cannot make anything useful of it. To do so you need to use the Text type, which is able to properly output text with several options such as character size, style, color, position, rotation, etc. This separation allows more flexibility and better performances: indeed a Font is a heavy resource, and any operation on it is slow (often too slow for real-time applications). On the other side, a Text is a lightweight object which can combine the glyphs data and metrics of a Font to display any text on a render target. Note that it is also possible to bind several Text instances to the same Font.

It is important to note that the Text instance doesn't copy the font that it uses, it only keeps a reference to it. Thus, a Font must not be destructed while it is used by a Text (i.e. never write a function that uses a local Font instance for creating a text).

Apart from loading font files, and passing them to instances of Text, you should normally not have to deal directly with this type. However, it may be useful to access the font metrics or rasterized glyphs for advanced usage.

Note that if the font is a bitmap font, it is not scalable, thus not all requested sizes will be available to use. This needs to be taken into consideration when using Text. If you need to display text of a certain size, make sure the corresponding bitmap font that supports that size is used.

Implementations

impl Font

pub fn kerning(&self, first: u32, second: u32, character_size: u32) -> i32

Get the kerning value corresponding to a given pair of characters in a font

Arguments

first - Unicode code point of the first character

second - Unicode code point of the second character

characterSize - Character size, in pixels

Return the kerning offset, in pixels

pub fn line_spacing(&self, character_size: u32) -> i32

Get the line spacing value

Arguments

characterSize - Character size, in pixels

Return the line spacing, in pixels

pub fn glyph(
&self,
codepoint: u32,
character_size: u32,
bold: bool,
outline_thickness: f32
) -> Glyph

Get a glyph in a font

Arguments

codePoint - Unicode code point of the character to get

characterSize - Character size, in pixels

bold - Retrieve the bold version or the regular one?

Return the corresponding glyph

pub fn info(&self) -> Info

Returns the font information.

pub fn underline_position(&self, character_size: u32) -> f32

Returns the position of the underline.

pub fn underline_thickness(&self, character_size: u32) -> f32

Returns the thickness of the underline.

pub fn from_file(filename: &str) -> Option<SfBox<Self>>

Load the font from a file.

The supported font formats are: TrueType, Type 1, CFF, OpenType, SFNT, X11 PCF, Windows FNT, BDF, PFR and Type 42. Note that this function know nothing about the standard fonts installed on the user's system, thus you can't load them directly.

Warning

SFML cannot preload all the font data in this function, so the file has to remain accessible until the Font object loads a new font or is destroyed.

pub fn from_stream<T: Read + Seek>(stream: &mut T) -> Option<SfBox<Self>>

Create a new font from a stream (a struct implementing Read and Seek)

Arguments

stream - Your struct, implementing Read and Seek

Returns None on failure.

pub fn from_memory(memory: &[u8]) -> Option<SfBox<Self>>

Create a new font from memory

Arguments

memory - The in-memory font file

Returns None on failure.

pub fn texture(&mut self, character_size: u32) -> &Texture

Get the texture containing the glyphs of a given size in a font

Arguments

characterSize - Character size, in pixels

Return the texture

Note: Unfortunately, this method requires mutable access, because CSFML uses a texture cache or something that it must update every time this function is called.