Struct sfml::graphics::RenderTexture

pub struct RenderTexture { /* fields omitted */ }

Target for off-screen 2D rendering into a texture

Implementations

impl RenderTexture

pub fn new(width: u32, height: u32, depth_buffer: bool) -> Option<RenderTexture>

Construct a new render texture

Arguments

width - Width of the render texture

height - Height of the render texture

depthBuffer - Do you want a depth-buffer attached? (useful only if you're doing 3D OpenGL on the rendertexture)

Return Some(RenderTexture) or None

pub fn display(&self)

Update the contents of the target texture

pub fn set_active(&mut self, active: bool) -> bool

Activate or deactivate a render texture as the current target for rendering

Arguments

active - true to activate, false to deactivate

pub fn texture(&self) -> &Texture

Get the target texture of a render texture

Return the target texture

pub fn set_smooth(&mut self, smooth: bool)

Enable or disable the smooth filter on a render texture

Arguments

smooth - true to enable smoothing, false to disable it

pub fn is_smooth(&self) -> bool

Tell whether the smooth filter is enabled or not for a render texture

Return true if smoothing is enabled, false if it is disabled

pub fn set_repeated(&mut self, repeated: bool)

Enable or disable texture repeating.

This function is similar to Texture::setRepeated. This parameter is disabled by default.

pub fn is_repeated(&self) -> bool

Tell whether the texture is repeated or not.

pub unsafe fn generate_mipmap(&mut self) -> bool

Generate a mipmap using the current texture data.

This function is similar to Texture::generate_mipmap and operates on the texture used as the target for drawing.

Safety

Be aware that any draw operation may modify the base level image data. For this reason, calling this function only makes sense after all drawing is completed and display has been called. Not calling display after subsequent drawing will lead to undefined behavior if a mipmap had been previously generated.