Module sfml::system

Base module of SFML, defining various utilities.

It provides vector types, timing types.

Structs

Clock
Utility type that measures the elapsed time.

SfBox
An owning pointer to an SFML-allocated object.

SfStr
A borrowed string type that's compatible with sf::String.

Time
Represents a time value.

Vector2
Utility type for manipulating 2-dimensional vectors.

Vector3
Utility type for manipulating 3-dimensional vectors.

Constants

FALSE
Boolean false value used by CSFML.

TRUE
Boolean true value used by CSFML.

Traits

SfStrConv
Trait for types that can be converted into SfStr

Functions

sleep
Make the current thread sleep for a given duration.

Type Definitions

Bool
Boolean type used by CSFML.

Vector2f
Vector2 with f32 coordinates.

Vector2i
Vector2 with i32 coordinates.

Vector2u
Vector2 with u32 coordinates.

Vector3f
Vector3 with f32 coordinates.

Vector3i
Vector3 with i32 coordinates.