Struct sfml::system::Time

pub struct Time(_);

Represents a time value.

Time encapsulates a time value in a flexible way.

It allows to define a time value either as a number of seconds, milliseconds or microseconds. It also works the other way round: you can read a time value as either a number of seconds, milliseconds or microseconds.

By using such a flexible interface, the API doesn't impose any fixed type or resolution for time values, and let the user choose its own favorite representation.

Time values support the usual mathematical operations: you can add or subtract two times, multiply or divide a time by a number, compare two times, etc.

Since they represent a time span and not an absolute time value, times can also be negative.

Usage example

let t1 = Time::seconds(0.1);
assert_eq!(t1.as_milliseconds(), 100);

let t2 = Time::milliseconds(30);
assert_eq!(t2.as_microseconds(), 30_000);

let t3 = Time::microseconds(-800_000);
assert_eq!(t3.as_seconds(), -0.8);

See also

Clock

Implementations

impl Time

pub fn seconds(seconds: f32) -> Self

Constructs a time value from a number of seconds.

pub fn milliseconds(milliseconds: i32) -> Self

Constructs a time value from a number of milliseconds.

pub fn microseconds(microseconds: i64) -> Self

Constructs a time value from a number of microseconds.

pub fn as_seconds(self) -> f32

Returns the time value as a number of seconds.

pub fn as_milliseconds(self) -> i32

Returns the time value as a number of milliseconds.

pub fn as_microseconds(self) -> i64

Returns the time value as a number of microseconds.

pub const ZERO: Time

Predefined "zero" time value.