Module sfml::window::sensor

Access to the real-time state of the sensors.

sensor provides an interface to the state of the various sensors that a device provides.

This module allows users to query the sensors values at any time and directly, without having to deal with a window and its events. Compared to the SensorChanged event, sensor can retrieve the state of a sensor at any time (you don't need to store and update its current value on your side).

Depending on the OS and hardware of the device (phone, tablet, ...), some sensor types may not be available. You should always check the availability of a sensor before trying to read it, with the is_available function.

You may wonder why some sensor types look so similar, for example Accelerometer and Gravity / UserAcceleration. The first one is the raw measurement of the acceleration, and takes into account both the earth gravity and the user movement. The others are more precise: they provide these components separately, which is usually more useful. In fact they are not direct sensors, they are computed internally based on the raw acceleration and other sensors. This is exactly the same for Gyroscope vs Orientation.

Because sensors consume a non-negligible amount of current, they are all disabled by default. You must call set_enabled for each sensor in which you are interested.

Usage example

use sfml::window::sensor;

if sensor::is_available(sensor::Type::Gravity) {
    // enable the gravity sensor
    sensor::set_enabled(sensor::Type::Gravity, true);
    // get the current value of gravity
    let _gravity = sensor::value(sensor::Type::Gravity);
}

Enums

Type
Sensor type.

Functions

is_available
Check if a sensor is available on the underlying platform.

set_enabled
Enable or disable a sensor.

value
Get the current sensor value.