Module sfml::window::touch

Access to the real-time state of the touches.

Touch provides an interface to the state of the touches.

This module allows users to query the touches state at any time and directly, without having to deal with a window and its events. Compared to the TouchBegan, TouchMoved and TouchEnded events, touch can retrieve the state of the touches at any time (you don't need to store and update a boolean on your side in order to know if a touch is down), and you always get the real state of the touches, even if they happen when your window is out of focus and no event is triggered.

There are two functions that can be used to retrieve the current position of a touch.

desktop_position: operates in global coordinates (relative to the desktop)

crate::window::Window::touch_position: operates in window coordinates (relative to a specific window).

Touches are identified by an index (the "finger"), so that in multi-touch events, individual touches can be tracked correctly. As long as a finger touches the screen, it will keep the same index even if other fingers start or stop touching the screen in the meantime. As a consequence, active touch indices may not always be sequential (i.e. touch number 0 may be released while touch number 1 is still down).

Usage example

use sfml::window::touch;

if touch::is_down(0) {
    // touch 0 is down
}

// get global position of touch 1
let _global_pos = touch::desktop_position(1);

// get position of touch 1 relative to a window
let _relative_pos = window.touch_position(1);

Functions

desktop_position
Get the current position of a touch in desktop coordinates.

is_down
Check if a touch event is currently down.